requirements-dev.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
chardet==3.0.4
click==7.1.1
entrypoints==0.3
flake8==3.7.9
gitdb==4.0.4
GitPython==3.1.1
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
jsonschema==3.2.0
mccabe==0.6.1
more-itertools==8.2.0
msgpack==1.0.0
multidict==4.7.5
packaging==20.3
pathspec==0.8.0
pluggy==0.13.1
pop==12
pop-config==5.5
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
pytest-asyncio==0.10.0
PyYAML==5.3.1
regex==2020.4.4
six==1.14.0
smmap==3.0.2
toml==0.10.0
typed-ast==1.4.1
wcwidth==0.1.9
yarl==1.4.2
zipp==3.1.0
